refactor: 조작 관련 클래스를 control로 묶어주기
