+++ src/control/index.js
@@ -1,0 +1,4 @@
+const BridgeGame = require('./BridgeGame');
+const BridgeGameControl = require('./BridgeGameControl');
+
+module.exports = { BridgeGame, BridgeGameControl };
